+++ bin/run-core-tests.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+set -eo pipefail
+
+cd `dirname $BASH_SOURCE`/../packages/core
+
+npm run test:nolint
+