+++ packages/collector/src/metrics/keywords.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var applicationUnderMonitoring = require('../util/applicationUnderMonitoring');
+var applicationUnderMonitoring = require('@instana/core').util.applicationUnderMonitoring;
 var logger;
 logger = require('../logger').getLogger('metrics/keywords', function(newLogger) {
   logger = newLogger;