+++ packages/core/src/util/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 module.exports = exports = {
+  applicationUnderMonitoring: require('./applicationUnderMonitoring'),
   atMostOnce: require('./atMostOnce'),
   buffer: require('./buffer'),
   clone: require('./clone'),